requirements.txt
python-dotenv
yfinance
matplotlib
seaborn
pyportfolioopt
python-binance
ta
TA-Lib
pandas
numpy
scikit-learn